fix: Deep copy k8s service config to get a real copy (#84)
The prod service image was overwritten with the dev service image during
flow creation. The k8s service config struct differs from the docker
compose service config struct in the sense that it contains slices so it
requires to be deep copied to create a copy with nothing shared with the
original struct.